External vRouter Plugin (#23)
* [External vRouter Plugin] Add support for external vrouter datapath.

* [External vRouter Plugin] Code cleanup for external vRouter plugin

Cleanup of code to add support for external vRouter datapath
plugins.

* [External vRouter Plugin] Support  for external vRouter plugin

Add support for external vRouter datapath plugins.

Main idea here is to split apt-install and install_dkms into
different methods and allow something to be called between.

Relation-set is used to parse info between the two charms, a
vrouter-plugin interface is added to the contrail-agent charm to
allow this. The contrail-agent executes the apt-install part of the
install script and then sets a parameter "agent-installed", this is
passed to the external plugin charm which triggers the external
plugin, and once finished a parameter "vrouter-plugged", is set and
passed to contrail-agent. The agent charm is then free to complete
the installation by running install_dkms.

Signed-off-by: Pieter Malan <pieter.malan@netronome.com>

* [External vRouter Plugin] Remove print statements

* rework plugin mechanism

* cleanup update-status

* [External vRouter Plugin] Add support for external vRouter plugin.

This commit adds changes to the charm as proposed by Andrey.

* [External vRouter Plugin]Replace variable with existing variable

Replace lazy-intalled variable with vhost-ready

Signed-of-by: Pieter Malan <pieter.malan@netronome.com>

* fix init steps sequence in install

* [External vRouter Plugin] Integrate Andrey's changes.

Integrate Andrey's changes in the install sequence, and add
lazy_install to vrouter-plugin-changed

Signed-of-by: Pieter Malan <pieter.malan@netronome.com>