Allow using /etc/hosts for RabbitMQ name resolution (#51)
* Allow using /etc/hosts for rmq name resolution

In multi-homed setups interface names may be resolved to FQDNs with the
following scheme:

<interface_name>.<hostname>.<domain>

For example for "secondary" interfaces in MAAS the resolutions will be
as follows for bare metal hosts, LXD containers and VMs respectively:

1) bond3.control-3.maas;
2) eth2.juju-720d5c-1-lxd-3.maas;
3) ens4.juju-720d5c-1-kvm-3.maas.

If shortname derivation logic takes the first part of an FQDN for this
purpose, <interface_name> becomes a short name which is an incorrect
name that does not exist in the DNS system used by a host. As a result,
rabbitmq deployed into a docker container will fail to start as cluster
node names are not possible to resolve.

Although FQDNs are possible to use as of rabbitmq 3.4.0, there is
currently no support for that in automation that runs in the rabbitmq
docker image for Contrail.

This change adds a way to manage hostnames for rabbitmq via /etc/hosts
similarly to how it is done in charm-rabbitmq.

Hostnames are generated based on a hostname used in the system (or a
UTS namespace). A prefix '-contrail-rmq' is added to the hostname to
avoid name clashes with names present in the DNS used by a host to which
the rabbitmq docker container is deployed.

Hostname synchronization is done over a peer relation. Addresses are
selected either based on the 'control-network' config parameter.

* Fix issues with the previous commit

* typos fixed after functional testing;
* handle the lack of -changed events for peer units that come up later
  than others;
* handle placing a hostname into a peer relation bucket on
  control-network config changes.

* Additional fixes

* do not write the same hostname to /etc/hosts multiple times;
* flush unitdata to persist the changes;
* add more logging to controller-cluster-relation-{joined,changed};
* do not exit prematurely while handling cluster-relation-changed -
  a leader check prevented the hostname update code from running.